Added function m_primefac.c and documentation.

+++ src/math/m_log.c
@@ -6,13 +6,13 @@
 /*   By: jkoskela <jkoskela@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/02 16:26:55 by jkoskela          #+#    #+#             */
-/*   Updated: 2020/10/20 02:52:17 by jkoskela         ###   ########.fr       */
+/*   Updated: 2020/10/20 03:57:11 by jkoskela         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../libft.h"
 
-int					m_log(uint64_t n)
+int64_t					m_log(int64_t n)
 {
 	static uint64_t	x = 0x07EDD5E59A4E28C2;
 	static int		tab64[64] = {
